add more to window/common.py
